[FIX] Появление scroll bar при ошибках в регистрации и авторизации. Отображение полоски опыта игрока в лобби.
