+++ client/src/components/UI/index.ts
@@ -1,5 +1,4 @@
 import Button from "./Button/Button";
 import Input from "./Input/Input";
 import Logo from "./Logo/Logo";
-import { Alert } from "./Alert/Alert";
-export { Logo, Input, Button, Alert };
+export { Logo, Input, Button };
