+++ client/src/components/index.ts
@@ -1,6 +1,5 @@
-export { Dossier } from "./Dossier/Dossier";
 export { ProgressBar, CardAdvise } from "./Loading";
-export { Logo, Button, Input, Alert } from "./UI";
+export { Logo, Button, Input } from "./UI";
 export { Chat } from "./Chat/Chat";
 export { Modal } from "./Modal/Modal";
 export { AppRouter } from "./AppRouter/AppRouter";
