+++ client/src/pages/AuthPages/AuthPage.css
@@ -7,11 +7,11 @@
 
 .auth_title {
    width: fit-content;
-   padding: 20px 50px;
+   padding: 10px 50px;
    background-color: var(--dark-green);
    border-radius: 60px;
    border: 5px solid var(--black);
-   margin-bottom: 50px;
+   margin-bottom: 45px;
    margin-top: 5px;
 
    text-align: center;
@@ -81,15 +81,29 @@
    }
 }
 
-@media (max-width: 1130px) {
-   .errors_div div {
-      font-size: 45px;
-      border: 4px solid #000;
+@media (max-width:1400px)  {
+   .auth_title {
+      margin-bottom: 20px;
+      font-size: 55px;
+      padding: 10px 30px;
    }
 
    .auth_footer button {
-      font-size: 45px;
+      font-size: 50px;
       padding: 0 10px;
+   }
+
+   .errors_div div {
+      font-size: 45px;
+   }
+}
+
+@media (max-width: 1130px) {
+   .errors_div div {
+      border: 4px solid #000;
+   }
+   .auth_footer button {
+      font-size: 45px;
    }
 }
